+++ tests_integration/markets/omen/test_local_chain.py
@@ -3,9 +3,9 @@ from web3 import Web3
 
 from prediction_market_agent_tooling.gtypes import xDai
 from prediction_market_agent_tooling.tools.balances import get_balances
-from prediction_market_agent_tooling.tools.web3_utils import send_xdai_to, xdai_to_wei
+from prediction_market_agent_tooling.tools.web3_utils import xdai_to_wei
 from tests_integration.conftest import local_web3_at_block
-from tests_integration.local_chain_utils import get_anvil_test_accounts
+from tests_integration.local_chain_utils import get_anvil_test_accounts, send_xdai_to
 
 
 def test_connect_local_chain(local_web3: Web3) -> None:
